Move ValveSchedule to its own file
