+++ src/peripherals/valve/ValveConfig.hpp
@@ -8,6 +8,7 @@
 #include <kernel/Log.hpp>
 
 #include <peripherals/valve/ValveComponent.hpp>
+#include <peripherals/valve/ValveSchedule.hpp>
 #include <peripherals/valve/ValveScheduler.hpp>
 
 using namespace std::chrono;
@@ -87,40 +88,3 @@ void convertFromJson(JsonVariantConst src, ValveControlStrategyType& dst) {
 }
 
 }    // namespace farmhub::peripherals::valve
-
-namespace ArduinoJson {
-
-using farmhub::peripherals::valve::ValveSchedule;
-template <>
-struct Converter<ValveSchedule> {
-    static void toJson(const ValveSchedule& src, JsonVariant dst) {
-        JsonObject obj = dst.to<JsonObject>();
-        auto startLocalTime = src.getStart();
-        auto startTime = system_clock::to_time_t(startLocalTime);
-        tm startTm;
-        localtime_r(&startTime, &startTm);
-        char buf[64];
-        strftime(buf, sizeof(buf), "%FT%TZ", &startTm);
-        obj["start"] = buf;
-        obj["period"] = src.getPeriod().count();
-        obj["duration"] = src.getDuration().count();
-    }
-
-    static ValveSchedule fromJson(JsonVariantConst src) {
-        tm startTm;
-        strptime(src["start"].as<const char*>(), "%FT%TZ", &startTm);
-        auto startTime = mktime(&startTm);
-        auto startLocalTime = system_clock::from_time_t(startTime);
-        seconds period = seconds(src["period"].as<int>());
-        seconds duration = seconds(src["duration"].as<int>());
-        return ValveSchedule(startLocalTime, period, duration);
-    }
-
-    static bool checkJson(JsonVariantConst src) {
-        return src["start"].is<const char*>()
-            && src["period"].is<int>()
-            && src["duration"].is<int>();
-    }
-};
-
-}    // namespace ArduinoJson
