+++ src/peripherals/valve/ValveScheduler.hpp
@@ -5,6 +5,7 @@
 #include <list>
 
 #include <kernel/Time.hpp>
+#include <peripherals/valve/ValveSchedule.hpp>
 
 using namespace std::chrono;
 using namespace farmhub::kernel;
@@ -25,35 +26,6 @@ bool convertToJson(const ValveState& src, JsonVariant dst) {
 void convertFromJson(JsonVariantConst src, ValveState& dst) {
     dst = static_cast<ValveState>(src.as<int>());
 }
-
-class ValveSchedule {
-public:
-    ValveSchedule(
-        time_point<system_clock> start,
-        seconds period,
-        seconds duration)
-        : start(start)
-        , period(period)
-        , duration(duration) {
-    }
-
-    time_point<system_clock> getStart() const {
-        return start;
-    }
-
-    seconds getPeriod() const {
-        return period;
-    }
-
-    seconds getDuration() const {
-        return duration;
-    }
-
-private:
-    time_point<system_clock> start;
-    seconds period;
-    seconds duration;
-};
 
 struct ValveStateUpdate {
     ValveState state;
